Updated helm chart for use with v0.0.4 of vmss-prototype
This adds the minimum image age feature to the helm chart such
that the container version 0.0.4+ will use it.  Note that this
means that this helm chart must be used with 0.0.4 or later
containers due to providing that extra command line option.

This also fixes the "dryRun" setting in the helm chart to actually
be wired to the --dry-run option of the container.  (I had forgotten
to complete that last time)

Updated smoketest3.sh to be dry-run and set an image age